test: consolidate QA-196/QA-162/QA-195 into one shared-instance suite
relationship-parent-delete-cascade, relationship-transaction-atomicity, and
resource-status-contract each booted their own vanilla-default Harper
instance. Their fixtures are schema-compatible (identical Order/OrderItem
schema in the two relationship tests, a disjoint Kv table in the status-
contract test, no route/table name collisions), so merge them into one
relationship-and-status-contract suite that boots a single shared instance,
grouped by originating QA-id with a consolidated fixture (schema.graphql +
resources.js). Cuts the integration-test boot count for this scenario group
from 3 to 1.
