+++ integrationTests/resources/relationship-and-status-contract/config.yaml
