+++ integrationTests/resources/relationship-and-status-contract/schema.graphql
@@ -1,0 +1,30 @@
+# Merged fixture for QA-196, QA-162, QA-195 — shared Harper instance
+# (instance-affinity consolidation; see relationship-and-status-contract.test.ts).
+#
+# Order/OrderItem: identical schema used by QA-196 (parent-DELETE cascade) and
+# QA-162 (cross-table transaction atomicity) — no namespacing needed, both
+# fixtures defined this schema verbatim.
+#
+# Relationship model:
+#   Order.items      = @relationship(to: orderId)   reverse (parent -> children list)
+#   OrderItem.order  = @relationship(from: orderId)  forward (child -> parent)
+
+type Order @table @export {
+	id: ID @primaryKey
+	total: Float
+	items: [OrderItem] @relationship(to: orderId)
+}
+
+type OrderItem @table @export {
+	id: ID @primaryKey
+	orderId: ID @indexed
+	name: String
+	price: Float
+	order: Order @relationship(from: orderId)
+}
+
+# Kv: backing table for QA-195's custom-resource status/throw contract probes.
+type Kv @table @export {
+	id: ID @primaryKey
+	v: String
+}
