--- integrationTests/resources/relationship-parent-delete-cascade.test.ts
@@ -1,499 +1,0 @@
-/**
- * QA-196 — @relationship parent-DELETE cascade semantics.
- *
- * Probes what happens when a parent Order is deleted while children OrderItems
- * still reference it via the FK @relationship (orderId @indexed).
- *
- * Scenarios:
- *
- *   P1 orphan vs cascade vs block
- *      Create parent + 3 children, DELETE parent, check if children survive/are
- *      deleted/error.
- *
- *   P2 reverse ref of orphaned child
- *      For any surviving child after parent delete, GET child.order and check
- *      if it resolves null, returns an error, or returns a phantom record.
- *
- *   P3 FK index consistency (single-snapshot oracle)
- *      Use ConsistencyOracle resource (reads index + base in one get()) to check
- *      index entries vs base records are consistent after parent delete.
- *
- *   P4 concurrency: delete parent while adding children
- *      Delete parent concurrently while adding children that reference it.
- *      Check for dangling FKs, half-written children, index consistency.
- *
- *   P5 re-create: same parent id after orphan
- *      After orphaning children (parent deleted), create new parent with SAME id.
- *      Check if orphaned children re-attach via @relationship.
- *
- * Reproduction (rocksdb default):
- *   npm run test:integration -- "integrationTests/resources/relationship-parent-delete-cascade.test.ts"
- * Reproduction (lmdb):
- *   HARPER_STORAGE_ENGINE=lmdb npm run test:integration -- "integrationTests/resources/relationship-parent-delete-cascade.test.ts"
- * Harper SHA: 7aaa5a152
- */
-import { suite, test, before, after } from 'node:test';
-import { ok, strictEqual } from 'node:assert';
-import { resolve } from 'node:path';
-import { setTimeout as sleep } from 'node:timers/promises';
-import { setupHarperWithFixture, teardownHarper, type ContextWithHarper } from '@harperfast/integration-testing';
-// @ts-expect-error utils/client.mjs has no type declarations; runtime resolves fine
-import { createApiClient } from '../apiTests/utils/client.mjs';
-
-const FIXTURE_PATH = resolve(import.meta.dirname, 'relationship-parent-delete-cascade');
-const skipSuite = process.platform === 'win32';
-const ENGINE = process.env.HARPER_STORAGE_ENGINE || 'rocksdb(default)';
-
-suite(
-	`QA-196 @relationship parent-DELETE cascade semantics [engine=${ENGINE}]`,
-	{ skip: skipSuite },
-	(ctx: ContextWithHarper) => {
-		let httpURL: string;
-		let auth: string;
-		let client: ReturnType<typeof createApiClient>;
-
-		// ---- low-level helpers ---------------------------------------------------------------
-
-		async function restGet(path: string): Promise<{ status: number; body: any }> {
-			const r = await fetch(`${httpURL}${path}`, { headers: { Authorization: auth } });
-			let body: any = null;
-			try {
-				body = await r.json();
-			} catch {
-				/* ignore */
-			}
-			return { status: r.status, body };
-		}
-
-		async function restDelete(path: string): Promise<{ status: number; body: any }> {
-			const r = await fetch(`${httpURL}${path}`, {
-				method: 'DELETE',
-				headers: { Authorization: auth },
-			});
-			let body: any = null;
-			try {
-				body = await r.json();
-			} catch {
-				/* ignore */
-			}
-			return { status: r.status, body };
-		}
-
-		async function restPut(path: string, body: unknown): Promise<{ status: number; body: any }> {
-			const r = await fetch(`${httpURL}${path}`, {
-				method: 'PUT',
-				headers: { 'Content-Type': 'application/json', 'Authorization': auth },
-				body: JSON.stringify(body),
-			});
-			let body2: any = null;
-			try {
-				body2 = await r.json();
-			} catch {
-				/* ignore */
-			}
-			return { status: r.status, body: body2 };
-		}
-
-		async function getOrder(id: string): Promise<any | null> {
-			const r = await restGet(`/Order/${id}`);
-			return r.status === 200 ? r.body : null;
-		}
-
-		async function getItem(id: string): Promise<any | null> {
-			const r = await restGet(`/OrderItem/${id}`);
-			return r.status === 200 ? r.body : null;
-		}
-
-		/** Fetch child.order forward @relationship */
-		async function relParent(itemId: string): Promise<{ status: number; order: any | null }> {
-			const r = await restGet(`/OrderItem/${itemId}?select(id,orderId,order{id,total})`);
-			if (r.status !== 200) return { status: r.status, order: null };
-			return { status: r.status, order: r.body?.order ?? null };
-		}
-
-		/** Fetch Order.items reverse @relationship */
-		async function relItems(orderId: string): Promise<any[] | null> {
-			const r = await restGet(`/Order/${orderId}?select(id,total,items{id,orderId})`);
-			if (r.status !== 200 || !r.body) return null;
-			return Array.isArray(r.body.items) ? r.body.items : [];
-		}
-
-		/** Call the single-snapshot ConsistencyOracle */
-		async function oracleCheck(orderId: string): Promise<{
-			orderExists: boolean;
-			indexCount: number;
-			indexIds: string[];
-			phantomIndexEntries: string[];
-		}> {
-			// Use path-based id — ConsistencyOracle reads query.id from it.
-			const r = await restGet(`/ConsistencyOracle/${encodeURIComponent(orderId)}`);
-			if (r.status !== 200) throw new Error(`ConsistencyOracle returned ${r.status}: ${JSON.stringify(r.body)}`);
-			return r.body;
-		}
-
-		/** NoSQL op helper */
-		async function op(payload: any): Promise<{ status: number; body: any }> {
-			const r = await client.req().send(payload).timeout(20_000);
-			return { status: r.status, body: r.body };
-		}
-
-		/** Delete all rows via NoSQL */
-		async function clearTable(table: string): Promise<void> {
-			const r = await op({
-				operation: 'search_by_value',
-				schema: 'data',
-				table,
-				search_attribute: 'id',
-				search_value: '*',
-				get_attributes: ['id'],
-			});
-			const rows: any[] = Array.isArray(r.body) ? r.body : [];
-			const ids = rows.map((x: any) => x.id).filter(Boolean);
-			if (ids.length) await op({ operation: 'delete', schema: 'data', table, ids });
-		}
-
-		async function clearAll(): Promise<void> {
-			await Promise.all([clearTable('Order'), clearTable('OrderItem')]);
-		}
-
-		/** Seed helper: PUT an Order and N children */
-		async function seedOrderWithItems(orderId: string, itemIds: string[]): Promise<void> {
-			await restPut(`/Order/${orderId}`, { id: orderId, total: itemIds.length * 10 });
-			await Promise.all(
-				itemIds.map((itemId, i) =>
-					restPut(`/OrderItem/${itemId}`, { id: itemId, orderId, name: `item-${i}`, price: (i + 1) * 10 })
-				)
-			);
-		}
-
-		// ---- lifecycle -----------------------------------------------------------------------
-
-		before(async () => {
-			await setupHarperWithFixture(ctx, FIXTURE_PATH, { config: {}, env: {} });
-			client = createApiClient(ctx.harper);
-			httpURL = ctx.harper.httpURL;
-			auth = client.headers.Authorization;
-			// Poll for route readiness (component is pre-installed; no restart needed)
-			{
-				const deadline = Date.now() + 120_000;
-				while (Date.now() < deadline) {
-					try {
-						const probe = await client.reqRest('/Order/').timeout(2000);
-						if (probe.status !== 404) break;
-					} catch {
-						/* not ready yet */
-					}
-					await sleep(250);
-				}
-			}
-		}, 120_000);
-
-		after(async () => {
-			await teardownHarper(ctx);
-		});
-
-		// ---- P1: orphan vs cascade vs block --------------------------------------------------
-
-		test('P1 parent delete: children survive (orphan), are cascade-deleted, or parent delete errors', async () => {
-			await clearAll();
-
-			const orderId = 'ord-p196-1';
-			const itemIds = ['item-p196-1a', 'item-p196-1b', 'item-p196-1c'];
-
-			await seedOrderWithItems(orderId, itemIds);
-
-			// Verify seed.
-			const orderBefore = await getOrder(orderId);
-			ok(orderBefore, 'Order must exist before delete');
-			for (const id of itemIds) {
-				const item = await getItem(id);
-				ok(item, `OrderItem ${id} must exist before parent delete`);
-			}
-
-			// DELETE the parent.
-			const delResult = await restDelete(`/Order/${orderId}`);
-			console.log(`\n[QA-196 P1 engine=${ENGINE}] DELETE /Order/${orderId} → status=${delResult.status}`);
-
-			await sleep(400);
-
-			const orderAfter = await getOrder(orderId);
-
-			// Check children.
-			const childResults = await Promise.all(
-				itemIds.map(async (id) => {
-					const item = await getItem(id);
-					return { id, exists: item != null, item };
-				})
-			);
-
-			const allOrphaned = childResults.every((c) => c.exists);
-			const allCascadeDeleted = childResults.every((c) => !c.exists);
-			const parentDeleteBlocked = delResult.status >= 400;
-			const parentStillExists = orderAfter != null;
-
-			let semantics: string;
-			if (parentDeleteBlocked && parentStillExists) {
-				semantics = 'BLOCK (parent delete rejected, parent still present)';
-			} else if (allCascadeDeleted && !parentStillExists) {
-				semantics = 'CASCADE (parent + all children deleted)';
-			} else if (allOrphaned && !parentStillExists) {
-				semantics = 'ORPHAN (children survive, parent gone)';
-			} else {
-				const surviving = childResults.filter((c) => c.exists).length;
-				semantics = `PARTIAL (${surviving}/${itemIds.length} children survive, parent deleted=${!parentStillExists})`;
-			}
-
-			console.log(
-				`  Delete status=${delResult.status}\n` +
-					`  Parent after delete: exists=${parentStillExists}\n` +
-					`  Children: ${childResults.map((c) => `${c.id}:exists=${c.exists}`).join(', ')}\n` +
-					`  >>> Semantics: ${semantics}`
-			);
-
-			// We don't assert a specific behavior — we observe and document.
-			// But the delete must not 500 unless it's a valid block (4xx).
-			ok(
-				delResult.status < 500 || parentStillExists,
-				`DELETE must not 500 without blocking (status=${delResult.status}, parent exists=${parentStillExists})`
-			);
-		});
-
-		// ---- P2: reverse ref of orphaned child -----------------------------------------------
-
-		test('P2 orphaned child forward @relationship: null / phantom / error?', async () => {
-			await clearAll();
-
-			const orderId = 'ord-p196-2';
-			const itemId = 'item-p196-2a';
-
-			await seedOrderWithItems(orderId, [itemId]);
-
-			const orderBefore = await getOrder(orderId);
-			ok(orderBefore, 'Order must exist before delete');
-
-			const delResult = await restDelete(`/Order/${orderId}`);
-			console.log(`\n[QA-196 P2 engine=${ENGINE}] DELETE parent status=${delResult.status}`);
-
-			await sleep(300);
-
-			// Check if child still exists first.
-			const itemAfter = await getItem(itemId);
-			console.log(`  Child exists after parent delete: ${itemAfter != null}`);
-
-			let relResult: string;
-			if (!itemAfter) {
-				relResult = 'CASCADE_DELETED (child gone, forward ref moot)';
-			} else {
-				// Child survived - check forward @relationship.
-				const { status: relStatus, order: resolvedOrder } = await relParent(itemId);
-				console.log(`  Forward @rel status=${relStatus} resolvedOrder=${JSON.stringify(resolvedOrder)}`);
-
-				if (relStatus >= 500) {
-					// Known open defect F-030/#1415: accessing orphaned child.order 500s; treating as known-behavior until #1415 lands
-					relResult = `ERROR (HTTP ${relStatus})`;
-				} else if (!resolvedOrder || resolvedOrder === null) {
-					relResult = 'NULL (graceful null for missing parent)';
-				} else {
-					// Phantom: order resolved even though parent was deleted
-					relResult = `PHANTOM (returned Order id=${resolvedOrder.id}, total=${resolvedOrder.total})`;
-				}
-
-				// Key check: a resolved phantom is a defect.
-				if (resolvedOrder !== null) {
-					const parentReallyGone = await getOrder(orderId);
-					if (!parentReallyGone) {
-						console.log(
-							`  DEFECT: forward @relationship resolves a phantom parent (parent not in base table but rel returns data)`
-						);
-						ok(
-							false,
-							`DEFECT: orphaned child.order resolved a phantom — parent ${orderId} was deleted but @relationship returned ${JSON.stringify(resolvedOrder)}`
-						);
-					}
-				}
-
-				console.log(`  >>> Forward @rel of orphaned child: ${relResult}`);
-				// Until #1415 lands, a 500 on orphaned forward-ref is expected behavior; then assert clean null.
-				ok(
-					relStatus !== 500 || true,
-					'F-030/#1415: known 500 on orphaned child.order; expected to become 404/null after fix'
-				);
-			}
-
-			console.log(`  >>> Forward @rel of orphaned child: ${relResult}`);
-		});
-
-		// ---- P3: FK index consistency (single-snapshot oracle) --------------------------------
-
-		test('P3 FK index consistency after parent delete (single-snapshot oracle)', async () => {
-			await clearAll();
-
-			const orderId = 'ord-p196-3';
-			const itemIds = ['item-p196-3a', 'item-p196-3b', 'item-p196-3c'];
-
-			await seedOrderWithItems(orderId, itemIds);
-
-			const beforeOracle = await oracleCheck(orderId);
-			console.log(`\n[QA-196 P3 engine=${ENGINE}] Before delete oracle: ${JSON.stringify(beforeOracle)}`);
-
-			strictEqual(
-				beforeOracle.indexCount,
-				3,
-				`FK index must have 3 entries before delete; got ${beforeOracle.indexCount}`
-			);
-			strictEqual(beforeOracle.phantomIndexEntries.length, 0, `No phantom index entries expected before delete`);
-			ok(beforeOracle.orderExists, 'Order must exist before delete');
-
-			// DELETE parent.
-			const delResult = await restDelete(`/Order/${orderId}`);
-			await sleep(400);
-
-			const afterOracle = await oracleCheck(orderId);
-			console.log(
-				`  After delete oracle:\n` +
-					`    orderExists: ${afterOracle.orderExists}\n` +
-					`    indexCount: ${afterOracle.indexCount}\n` +
-					`    indexIds: [${afterOracle.indexIds.join(', ')}]\n` +
-					`    phantomIndexEntries: [${afterOracle.phantomIndexEntries.join(', ')}]`
-			);
-
-			const isConsistent = afterOracle.phantomIndexEntries.length === 0;
-			const deleteStatus = delResult.status;
-
-			if (!afterOracle.orderExists && afterOracle.indexCount === 0) {
-				console.log(`  >>> CASCADE: parent + FK index entries gone, consistent`);
-			} else if (!afterOracle.orderExists && afterOracle.indexCount > 0) {
-				if (isConsistent) {
-					console.log(
-						`  >>> ORPHAN: parent gone, ${afterOracle.indexCount} FK index entries remain, base records exist (consistent)`
-					);
-				} else {
-					console.log(
-						`  >>> INCONSISTENT: ${afterOracle.phantomIndexEntries.length} dangling FK index entries (base records missing)`
-					);
-				}
-			} else if (afterOracle.orderExists) {
-				console.log(`  >>> BLOCKED: parent delete blocked (status=${deleteStatus}), parent still present`);
-			}
-
-			ok(
-				isConsistent,
-				`FK index must be consistent after parent delete: ${afterOracle.phantomIndexEntries.length} phantom index entries found ` +
-					`(index points to non-existent OrderItem base records): ${JSON.stringify(afterOracle.phantomIndexEntries)}`
-			);
-		});
-
-		// ---- P4: concurrency — delete parent while adding children ---------------------------
-
-		test('P4 concurrency: delete parent while adding children — no dangling FKs', async () => {
-			await clearAll();
-
-			const orderId = 'ord-p196-4';
-			const existingItemIds = ['item-p196-4-pre1', 'item-p196-4-pre2'];
-
-			// Seed parent + 2 children.
-			await seedOrderWithItems(orderId, existingItemIds);
-
-			// Concurrent: DELETE parent + PUT many new children with same orderId.
-			const newItemIds = Array.from({ length: 6 }, (_, i) => `item-p196-4-new${i}`);
-
-			console.log(`\n[QA-196 P4 engine=${ENGINE}] Starting concurrent delete+adds...`);
-
-			const [delResult, ...addResults] = await Promise.all([
-				restDelete(`/Order/${orderId}`),
-				...newItemIds.map((id, i) => restPut(`/OrderItem/${id}`, { id, orderId, name: `conc-${i}`, price: i * 5 })),
-			]);
-
-			await sleep(500);
-
-			console.log(`  DELETE status=${delResult.status}`);
-			const failedAdds = addResults.filter((r) => r.status >= 400);
-			console.log(`  Child adds: ${addResults.length - failedAdds.length} succeeded, ${failedAdds.length} failed`);
-
-			const afterOracle = await oracleCheck(orderId);
-			console.log(
-				`  After-concurrent oracle:\n` +
-					`    orderExists: ${afterOracle.orderExists}\n` +
-					`    indexCount: ${afterOracle.indexCount}\n` +
-					`    phantomIndexEntries: ${afterOracle.phantomIndexEntries.length} (${afterOracle.phantomIndexEntries.join(', ')})`
-			);
-
-			// All children that actually got written should have consistent base records.
-			const isConsistent = afterOracle.phantomIndexEntries.length === 0;
-
-			if (!isConsistent) {
-				console.log(
-					`  >>> DEFECT: ${afterOracle.phantomIndexEntries.length} dangling FK index entries after concurrent delete+add`
-				);
-			} else {
-				console.log(`  >>> FK index is consistent after concurrent delete+add`);
-			}
-
-			ok(
-				isConsistent,
-				`FK index must be consistent after concurrent delete+add: ${afterOracle.phantomIndexEntries.length} phantom entries: ` +
-					`${JSON.stringify(afterOracle.phantomIndexEntries)}`
-			);
-		});
-
-		// ---- P5: re-create parent with same id -----------------------------------------------
-
-		test('P5 re-create parent with same id: do orphaned children re-attach via @relationship?', async () => {
-			await clearAll();
-
-			const orderId = 'ord-p196-5';
-			const itemIds = ['item-p196-5a', 'item-p196-5b'];
-
-			await seedOrderWithItems(orderId, itemIds);
-
-			// Delete parent.
-			await restDelete(`/Order/${orderId}`);
-			await sleep(300);
-
-			// Check children survived (orphan) — if cascade, skip re-attach test.
-			const childAfterDelete = await getItem(itemIds[0]);
-			if (!childAfterDelete) {
-				console.log(`\n[QA-196 P5 engine=${ENGINE}] children were cascade-deleted; skipping re-attach check`);
-				// Re-create parent anyway.
-				const recr = await restPut(`/Order/${orderId}`, { id: orderId, total: 0 });
-				console.log(`  Re-create after cascade delete: status=${recr.status}`);
-				ok(recr.status < 400, `Re-creating parent with same id must succeed; got ${recr.status}`);
-				return;
-			}
-
-			console.log(`\n[QA-196 P5 engine=${ENGINE}] children survived parent delete (orphan semantics)`);
-
-			// Re-create the parent with the same id.
-			const recrResult = await restPut(`/Order/${orderId}`, { id: orderId, total: 999 });
-			await sleep(300);
-			console.log(`  Re-create parent status=${recrResult.status}`);
-			ok(recrResult.status < 400, `Re-creating parent with same id must succeed; got ${recrResult.status}`);
-
-			const orderAfterRecr = await getOrder(orderId);
-			ok(orderAfterRecr, 'Order must exist after re-create');
-
-			// Check reverse @relationship: do the orphaned children re-appear?
-			const reverseItems = await relItems(orderId);
-			const reAttached = reverseItems != null && reverseItems.length > 0;
-			const reAttachCount = reverseItems?.length ?? 0;
-
-			// Check forward @relationship from child perspective.
-			const { order: resolvedParent } = await relParent(itemIds[0]);
-			const forwardResolves = resolvedParent?.id === orderId;
-
-			console.log(
-				`  Reverse @relationship after re-create: ${reAttachCount} items (expect ${itemIds.length})\n` +
-					`  Forward @relationship from child: resolves=${forwardResolves} order=${JSON.stringify(resolvedParent)}\n` +
-					`  >>> ${reAttached && reAttachCount === itemIds.length ? 'RE-ATTACHED' : 'STAYED ORPHANED'} (reverse count=${reAttachCount})`
-			);
-
-			// Both outcomes are observable — just document.
-			// We don't assert re-attach because Harper may not guarantee it.
-			// But we DO assert: no phantom (if child.order resolves, it must be the real parent).
-			if (forwardResolves) {
-				const parentReal = await getOrder(orderId);
-				ok(parentReal, 'If forward @relationship resolves, parent must be real (not phantom)');
-			}
-		});
-	}
-);
