--- integrationTests/resources/relationship-parent-delete-cascade/config.yaml
@@ -1,5 +1,0 @@
-graphqlSchema:
-  files: '*.graphql'
-jsResource:
-  files: resources/*.js
-rest: true
